test(architecture): calendar field-whitelist pairing + extract scan helper (Phase 1.5 #1686)
Adds CalendarFieldWhitelistPairingTests: every property a calendar surfaces
(StartPropertyName + optional End/Title/ColorBy) must appear in the matching
QueryDefinition<TEntity>.GetColumns() whitelist. The range-query endpoint
reads each event via the same projection — non-whitelisted properties either
render blank or leak data the security model intentionally hid. Same
hard-failure when an entity declares a calendar but no matching
QueryDefinition exists.

Refactors the assembly-scan boilerplate (~140 lines) out of the kanban
test into a shared EntityDefinitionScan helper:

- ScanEntityDefinitions / ScanQueryDefinitions / ReadColumnPropertyNames
- LoadGranitAssemblies / SafeGetTypes / InheritsFromOpenGeneric

KanbanCardWhitelistPairingTests now consumes the helper, dropping ~140 lines
of duplicated reflection code. The same helpers are ready for the upcoming
GalleryFieldWhitelistPairingTests (#1694).
